manuals/gb/ns.adoc: Update documentation regarding SNS capability
The SNS capability was added to our Gb stack quite some time ago,
but it seems the manual was not updated.  Let's update the document
accordingly.

The individual sub-sections about the SNS-* messages are empty as
I think lynxis is currently most qualified to fill them in, I'll
ask him to submit a follow-up patch.

Related: SYS#5212
Change-Id: I21e891fc7662b582681cd9bd7568e4b65d357751